// star-wars-api/src/startWars/enums/episodes.enum.ts
export enum EpisodesEnum {
    A_NEW_HOPE = 'A_NEW_HOPE',
    THE_EMPIRE_STRIKES_BACK = 'THE_EMPIRE_STRIKES_BACK',
    RETURN_OF_THE_JEDI = 'RETURN_OF_THE_JEDI',
    THE_PHANTOM_MENACE = 'THE_PHANTOM_MENACE',
    ATTACK_OF_THE_CLONES = 'ATTACK_OF_THE_CLONES',
    REVENGE_OF_THE_SITH = 'REVENGE_OF_THE_SITH',
    THE_FORCE_AWAKENS = 'THE_FORCE_AWAKENS',
    THE_LAST_JEDI = 'THE_LAST_JEDI',
    THE_RISE_OF_SKYWALKER = 'THE_RISE_OF_SKYWALKER'
}